+++ src/lasieve4/piii/psp.c
@@ -22,18 +22,18 @@
 #include <gmp.h>
 #include "lasieve-asm.h"
 
-extern unsigned long montgomery_inv_n;
-extern unsigned long *montgomery_modulo_n;
-extern unsigned long montgomery_modulo_R2[NMAX_ULONGS];
-extern unsigned long montgomery_ulongs;
+extern uint32_t montgomery_inv_n;
+extern uint32_t *montgomery_modulo_n;
+extern uint32_t montgomery_modulo_R2[NMAX_ULONGS];
+extern uint32_t montgomery_ulongs;
 
-extern void (*asm_mulmod) (unsigned long *, unsigned long *, unsigned long *);
-extern void (*asm_squmod) (unsigned long *, unsigned long *);
-extern void (*asm_add2) (unsigned long *, unsigned long *);
-extern void (*asm_diff) (unsigned long *, unsigned long *, unsigned long *);
+extern void (*asm_mulmod) (uint32_t *, uint32_t *, uint32_t *);
+extern void (*asm_squmod) (uint32_t *, uint32_t *);
+extern void (*asm_add2) (uint32_t *, uint32_t *);
+extern void (*asm_diff) (uint32_t *, uint32_t *, uint32_t *);
 
 #if 0
-int asm_cmp64(unsigned long * a, unsigned long * b)
+int asm_cmp64(uint32_t * a, uint32_t * b)
 {
   if (a[0] != b[0])
     return 1;
@@ -46,7 +46,7 @@ int asm_cmp64(unsigned long * a, unsigned long * b)
 /*******************************************************/
 int psp64()
 /*******************************************************/
-{ unsigned long x[2], ex[2], one[2], s;
+{ uint32_t x[2], ex[2], one[2], s;
   long e, i, b=0, v;
 
   if (!(montgomery_modulo_n[0] & 1))
@@ -116,7 +116,7 @@ int psp64()
 }
 
 #if 0
-int asm_cmp(unsigned long * a, unsigned long * b)
+int asm_cmp(uint32_t * a, uint32_t * b)
 {
   long i;
 
@@ -130,7 +130,7 @@ int asm_cmp(unsigned long * a, unsigned long * b)
 /*******************************************************/
 int psp(mpz_t n)
 /*******************************************************/
-{ unsigned long x[NMAX_ULONGS], ex[NMAX_ULONGS], one[NMAX_ULONGS], s;
+{ uint32_t x[NMAX_ULONGS], ex[NMAX_ULONGS], one[NMAX_ULONGS], s;
   long e, i, b, v;
 
   if (!set_montgomery_multiplication(n))
